Report the versions of dependencies.
This change updates the VersionServlet to no longer just report the version of the
main application artifact but also of all dependencies found in the classpath.
Dependencies are discovered by scanning the classpath and looking for
`pom.properties` files in `META-INF/maven` subdirectories.  These `pom.properties`
files are put into jars when Maven builds packages and includes the version of the
dependency making it ideal for this use case.